refactored character handler code
