+++ handlers/index.js
@@ -1,3 +1,4 @@
 exports.events = require ('./events.js');
 exports.help = require ('./help.js');
 exports.users = require ('./users.js');
+exports.characters = require ('./characters.js');
